POC [wip] (very!)
