+++ lib/rubocop/cop/team.rb
@@ -10,6 +10,8 @@ module RuboCop
     # first the ones needed for autocorrection (if any), then the rest
     # (unless autocorrections happened).
     class Team
+      using Base::Internals
+
       attr_reader :errors, :warnings, :updated_source_file, :cops
 
       alias updated_source_file? updated_source_file
